Add files via upload
Created a model predicting whether an Electric Vehicle passed the ‘Clean Alternative Fuel Vehicle Eligibility’ for which it can seen on the road. Achieving an accuracy score of 100% on Random Forest Classifier. Also done the visualization on Tableau, giving you the idea of EV make, model and their electric range.